fix issue where 16 years old is being returned
0 should be returned if cat is 0 years old
added an additional condition before the if/else if statements to check if the input value is "0". If it is, we set the catAge variable to 0. We also added quotes around the input values in the if/else if statements to ensure that they are treated as strings. This is important because the value of an input element is always a string, and we need to compare it to string values.